fix deps issue
