+++ .eslintrc.cjs
@@ -1,9 +1,22 @@
 module.exports = {
+	settings: {
+		react: {
+			version: 'detect',
+		},
+	},
 	env: {
 		browser: true,
-		es2021: true,
+		es2020: true,
 	},
-	extends: ['eslint:recommended', 'plugin:react/recommended', 'prettier'],
+	extends: [
+		'eslint:recommended',
+		'plugin:@typescript-eslint/recommended',
+		'plugin:react-hooks/recommended',
+		'plugin:react/recommended',
+		'prettier',
+	],
+	ignorePatterns: ['dist', '.eslintrc.cjs'],
+	parser: '@typescript-eslint/parser',
 	overrides: [
 		{
 			env: {
@@ -19,8 +32,12 @@ module.exports = {
 		ecmaVersion: 'latest',
 		sourceType: 'module',
 	},
-	plugins: ['react', 'prettier'],
+	plugins: ['react-refresh', 'react', 'prettier'],
 	rules: {
+		'react-refresh/only-export-components': [
+			'warn',
+			{ allowConstantExport: true },
+		],
 		'no-unused-vars': ['warn'],
 		'linebreak-style': ['error', 'unix'],
 		quotes: ['error', 'single'],
