+++ index.html
@@ -8,7 +8,7 @@
 	</head>
 	<body>
 		<div id="root"></div>
-		<script type="module" src="/src/main.jsx"></script>
+		<script type="module" src="/src/main.tsx"></script>
 		<!-- Cloudflare Web Analytics -->
 		<script
 			defer
