+++ package.json
@@ -5,9 +5,9 @@
 	"type": "module",
 	"scripts": {
 		"dev": "vite",
-		"build": "vite build",
+		"build": "tsc && vite build",
 		"preview": "vite preview",
-		"lint": "yarn prettier . --write"
+		"lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0"
 	},
 	"dependencies": {
 		"@arco-design/web-react": "^2.46.2",
@@ -29,6 +29,8 @@
 	"devDependencies": {
 		"@types/react": "^18.0.28",
 		"@types/react-dom": "^18.0.11",
+		"@typescript-eslint/eslint-plugin": "^6.3.0",
+		"@typescript-eslint/parser": "^6.3.0",
 		"@vitejs/plugin-react": "^3.1.0",
 		"autoprefixer": "^10.4.14",
 		"eslint": "^7.32.0 || ^8.2.0",
@@ -39,9 +41,11 @@
 		"eslint-plugin-prettier": "^5.0.0",
 		"eslint-plugin-react": "^7.33.1",
 		"eslint-plugin-react-hooks": "^4.3.0",
+		"eslint-plugin-react-refresh": "^0.4.3",
 		"postcss": "^8.4.27",
 		"prettier": "^3.0.0",
 		"tailwindcss": "^3.3.3",
+		"typescript": "^5.1.6",
 		"vite": "^4.2.0"
 	}
 }
