--- src/main.jsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import '@arco-design/web-react/dist/css/arco.css';
-import App from './App.jsx';
-
-ReactDOM.createRoot(document.getElementById('root')).render(<App />);
