+++ src/main.tsx
@@ -1,0 +1,7 @@
+import ReactDOM from 'react-dom/client';
+import '@arco-design/web-react/dist/css/arco.css';
+import App from './App.js';
+
+ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
+	<App />,
+);
