+++ tsconfig.node.json
@@ -1,9 +1,8 @@
 {
 	"compilerOptions": {
 		"composite": true,
-		"skipLibCheck": true,
 		"module": "ESNext",
-		"moduleResolution": "bundler",
+		"moduleResolution": "Node",
 		"allowSyntheticDefaultImports": true
 	},
 	"include": ["vite.config.ts"]
